👌(layout) customize CMS placeholder menu for more readability
CMS page placeholders menu put the plugin name and plugin object
title on the same line to save vertical space at the cost
of readability. It comes quickly to unreadable plugin labels and
leaded to shortenize plugin names that was not intuitive.

This commit hack the CMS placeholder menu CSS to change this
behavior to stacked plugin name and object title on their own line.
With some adjustments to preserve some vertical space it free some
horizontal space so the plugin name can display a long name and
object title can be longer and more accurate.