♻️(icons) change how the SVG icons sprite is used to not break CDNs
We previously determined that the best way to build our small icon set was
using SVGs with <use xlink:href> and <title> for accessibility. This worked
great throughout development but broke down as we tried to deploy it to
production environments: <use xlink:href> cannot load cross-domain external
resources.

We propose {% include %}-ing the svg sprite at the top of our base template
to avoid hitting that wall. This way we keep our icons entirely customizable
through CSS and still have access to the best way to include them:
<use xlink:href>. The icons are also way easier to access both by our backend
templates and our frontend components. It also provides a straightforward way
for richie users to override them: use Django's template overloads.

The only downside is that this does not take advantage of CDN caching. In our
opinion this is a small price to pay as our svg icon sprite is small in size
at ~5k.