🗃(models) add a model for enrollments in a new app
The Enrollment model links a user to the course runs they have enrolled in.

In most cases, actual enrollment also needs to happen in the actual LMS
behind Richie. Still, we need a simple and fast way to look up and manage
user enrollments in Richie itself, which is tough to achieve if every action
necessitates a bunch of API calls to different LMSes.

This is where the Enrollment model comes in: it is kept in sync with
enrollments in various LMSes but provides a unified Django way to know about
them and manage them.

We decided to start a new app where we put this model so as not to force
all Richie users to manage enrollments through Richie. This way, users who
do not add this new app to their INSTALLED_APPS will not have their schema
polluted by this new part of Richie.