✨(dropdown) add new layout component "dropdown"
A need for a dropdown component emerge from language menu which is
required to be a dropdown. Since we don't have any Framework
component available, we need to implement it first.

This commit adds this new component including the library "dropmic"
with a Javascript and Sass sources both modified to fit our
requirements (BEM names). A dropdown sample is available from
Styleguide page.

Since dropdown component has been added, we can integrate the language menu
with it and as required add it also in header.

This update some theme properties related from header and footer groups to
include their own CTA dropdown colors, update styleguide and language menu.