✨(api) open endpoints for enrollments
As we manage enrollments through Richie, we will need to be able
to create them as well as retrieve them from the frontend. To that
end, we're adding create/retrieve/list endpoints for enrollments
in our new enrollments app.

We chose to use a new viewset, including for enrollment creation, to
keep these features in the enrollments app and not create a dependency
on the enrollments app from the courses app.

As enrollments are private information belonging to one particular
user, we restricted the queryset on these endpoints to enrollments
owned by the current logged-in user.