💄(course) add icons to course glimpse footer & organization
Add a calendar icon to the course glimpse footer & a pin icon
next to the organization name in the glimpse.

This commit includes usage of these icons in course glimpse template
and adjust text aside in CSS to be correctly positioned. Also add two
new icons 'calendar' and 'pin' in SVG icon map file.