✅(frontend) add a Deferred util for frontend tests
We generally use fetchMock to mock backend API calls in our frontend
tests. This is practical but does not allow us to manage when exactly
the requests finish. It also does not hook into React internals to
manage act calls for us.

Using a Deferred, we can more finely control fetchMock calls'
resolution, allowing us to more easily test loading states and not run
afoul of React's act-related safeguards. It'll also be necessary when
testing-library finally removes the `wait()` helper.