✨(context) add flags for installed Richie apps on Richie's context
We need a convenient way to know if an optional Richie app is installed
or not. In our opinion, it's best to hook into Django's native way
and just look up "INSTALLED_APPS" to see what is actually installed or
not.

Additionally, we need to access this easily from templates to add or
remove UI elements depending on the installed apps. To that end, we
added the information through a RICHIE_APPS element on richie's context
processor.