✨(courses) integrate course run enrollment in course views
We need to integrate course run enrollments, through our new
<CourseRunEnrollment /> component, on the course detail view and the
course run detail view. We want to make enrollments seamless for users
but still give them links to the course run views even when there is
an enrollment button.

We're using the RICHIE_APPS.enrollments flag we just added to our
context processor to determine whether to include the enrollment
button or leave that to the LMS itself.