💄(course) improve search error message layout
We need a better error layout than the current simple text. It should
have a visual and text should be more visible.

Here we add a new icon related to search fail and update Search
component to include it. Then we adjust the CSS so icon and text are
accorded and nicer.