🔒(api) lock down not-in-use course run API endpoints
Building an API using DRF ViewSets is straightforward and
declarative, but it can lead to opening up endpoints or actions on
some objects unintentionally.

We decided to add comprehensive authorization testing for course
run endpoints to make sure there was no issue there. We were able
to close all unused endpoints by making a NotAllowed permission act
as a default for all non-specifically-specified endpoints on the
course runs ViewSet.