🔧(pylint) relax pylint's parents class count limitation
We're encounterin errors with the max-parent rule for classes
imported from dependencies such as Django Rest Framework. In our
opinion, this means the rule is too troublesome as it creates
unneeded warnings there while it has never raised a warning for
our own classes in Richie.

We can relax it so DRF classes can get under the limitation. It
would still catch egregious cases.