✅(frontend) refactor tests to stop using wait()
Because it is a somewhat non-defined "wait for a promise tick"
utility, wait() can lead to non-determined behavior in tests. As a
result, testing-library deprecated it and replaced it with waitFor()
which specifically waits for something we expect to happen and
processes async chores for us to get there.

Here, we use our new Deferred and waitFor() to get rid of all our
uses of wait(). This also removes all the deprecation warnings
that were being printed out during our tests.