🐛(frontend) fix the broken course search field after filters are reset
When we added tracking of the last dispatched action(s) to our state
setup, we refactored <SearchSuggestField /> to use that to determine
when to empty itself in reaction to "filterReset" actions, instead of
the previous approach of guessing based on a comparison between current
values and values during the previous render.

However, we did not take into account that leaving the check in the
body of the component, it would be executed on each render and prevent
any value from being entered after a "filterReset" action was issued.

We can solve it by placing the check & clear in a useEffect that only
runs when the last dispatched actions change. This way it will react
to incoming action by checking if it needs to empty itself, but not
even run the check if there is no new action and the user is just
interacting with the field.