✨(layout) add Linkedin icon and new section template "septenary"
Linkedin icon is required for social networks badges. Also we need a
new section template to include a new section background with a wave
decoration for the blogpost block on homepage.

This commit add the new icon in the SVG sprite map with its link
in social network templates, add a new wave decoration and a new
section template "septenary" to include the wave decoration to use
on blogpost block.