➕(frontend) add downshift & user-event to our dependencies
We're building a dropdown menu to allow users to select their
preferred language. As we already know we'll need another select
menu in Richie, which is already planned out (the user account
menu), we need a common foundation to build these two dropdown
menus from.

This is where downshift comes in: it gives us primitives to build
those menus without repeating ourselves and with as little effort
as possible.

As for user-event, it's a @testing-library helper that helps us
interact with UI elements by simulating what a real user would do.