♻️(oauth) stop synchronizing user permission fields on oauth
The is_staff and is_superuser were synchronized with the authentication
provider upon each login. This is not what we want as our teachers must
be staff in Richie for example, but not in edX. Being staff or superuser
on one side does not mean you should be on the other side.