✨(frontend) add a React-powered <LanguageSelector /> component
As we have a React codebase that powers all of Richie's interactive
interface elements, it makes sense for us to build the language
selector using our frontend stack.

We're relying on downshift to build a dropdown menu with less code,
and expecting the data to be passed directly from the backend template
as that is where Django-CMS makes it available.