✨(templates) integrate our language picker with our templates
Django-CMS allows us to integrate with its locale management
through the `language_chooser` template-tag. It allows us to use
our own template but does not do so through a template include.

This means we can integrate our React component in a template we
then supply to `language_chooser`, but it prevents us from passing
contextual data from the template include to the React component.
We're thus using CSS to target the two different pickers for the
footer and header.