💄(frontend) add a chevron icon to the language picker
Our language picker looks just like a link or a regular blob of text,
in bold. We need to give users a hint that they can click it to
open a menu. This is why we're adding a chevron icon to our html
sprite and placing it next to the text of the language picker button.