🔧(prettier) add default rules
As it is recommended in python to respect a line width of 79 characters,
there are different recommendations in js. I suggest to use the Airbnb standard
which recommends to use 100 characters per line.
As prettier also format html files, we must ignore django templates files
because some syntax do not support line break
but prettier automatically break lines.
Launch prettier to apply new rules on frontend and website dirs