🐛(frontend) use variables for language selector colors
When the language selector is displayed in the footer/header, it
must adapt to its surroundings by using appropriate colors. This
mostly goes for the text, which appears directly on the background
of the footer/header.

To achieve this, we can use variables from sass theming when defining
language selector colors.