💄(selector) add generic selector class
selector was previously language-selector. But we now use it also in
UserLinksSelector so it is convenient to create a generic selector class.
Also fix a little visual bug: apply display:none to the whole
language-selector__list--is-closed class not only to its :before & :after.
A little white point was visible on the screen when selector was displayed
on dark background.