✨(back) create a function checking username in social auth pipeline
When a username is already registered in richie database, python social
auth will prepend the username with a uuid to make it unique. We don't
want this behaviour because API call with edx don't work in this case,
we need the real username. We created a function, added in the python
social auth connection pipeline, responsible to check the username
availability and raise an exception if the same username already exists.