✨(back) extend social-auth middleware to use a page to display error
If login fails we want to display a customizable page. This page is a
CMS page, a reverse_id must be set on this page and added to the
SOCIAL_ERROR_REVERSE_ID setting. On a login error, if this setting is
defined then this page is used. It fallbacks on default behaviour
otherwise.