✨(richie/core) add registration oauth2 endpoint
Sign up url was the admin login page that is not correct
since richie use edx oauth2 for sso login. So we add a new
SOCIAL_AUTH_EDX_OAUTH2_REGISTRATION_ENDPOINT
setting variable to specify in settings the
endpoint to register on the oauth2 provider.
If no registration endpoint is specified,
we use the admin login page by default.