🔧(demo) allow configuring SITE_INFO pages
In some sites built with richie, we add for example a "contact us"
button that points to a page in the CMS. For development purposes,
the demo site must include a page with the reverse id chosen for
this page. This can only be done if we can add a page to the demo
at the level of the themed site.