🔧(build) fix the webpack babel regex for dependent builds
Dependent builds were broken: no loader was used to transform any
files from Richie proper. This was due to the node_module regex,
which prevented any file from richie used as a dependency to be
compiled.

We can explicitly exclude richie-education from our regexp to make
sure it is built as we intend.