📝(frontend) document dependent build overrides
Users of Richie who create their own dependent build probably
do so yo override some frontend components. As we're starting to
encounter this need, we thought it was high time to fill this gap
in our documentation and create some instructions for how to
achieve this through the mechanism we provided in Richie itself.