💄(frontend) add reboot style
Add a reboot style to apply generic style for the whole app. In particular, we
force html and body elements to take full width & height of the window.
It's useful to improve layout when content height is smaller than window height.
In this way, we can force main elements to take the full height of the document.
e.g: to avoid to have a footer with an ugly white banner below it.