✨(frontend) add UserLogout component
React component which takes logout endpoints in props then call them
through no-cors requests. Then wait that all requests are finished
and finally redirect to the uri provided through next query param if it is safe
or  to LOGOUT_REDIRECT_URL by default.

An uri is consider safe if its hostname is included
in REDIRECT_WHITELIST settings.If REDIRECT_WHITELIST is defined on [*], any
external uri is accepted. Conversely if value is [], any external uri
was consider as unsafe.