✨(back) add LimitBrowserCacheTTLHeaders middleware
By default, Django CMS send cached response with a
Cache-control:max-age value that reflects the server cache TTL.

The thing is : we can invalidate a server side cache entry, but we
cannot invalidate our client browser cache entries. That's why we want
to set a long TTL on the server side, but a much lower TTL on the
browser cache.

The LimitBrowserCacheTTLHeaders middleware and a new setting
(MAX_BROWSER_CACHE_TTL) have been introduced to handle this.