🚸(plugins) improve sanitizing of input text on CKEditor plugin
CKEditor is submitting text with web entities (e.g. &eacute;) and
potentially exotic characters. Our support team complained that the
count of the max number of characters was wrong. This commit fixes
it by normalizing and unescaping input strings. We do it for counting
and also before saving it to database.