🐛(frontend) refresh js on cms-content-refresh
In edit mode, when user saves a modification, the new content is hot replaced
but js scripts are not reinitiliazed. In our case, as we use React, all React
components are not rendered after a content modification in Edit mode.

To solve this issue, Django-CMS provides an event "cms-content-refresh". So we
listen this event and reinitialize our js each time this event is triggered.