♻️(courses) refactor course runs to not be CMS pages anymore
Course runs were initially built as CMS pages. It turns out that the
content they carry is structured and we do not benefit from their page
status. On the contrary it brings a lot of complexity that is getting
in our way to automate the synchronization of course runs with remote
LMSes.

I transformed the existing course run page extensions into simple
Django models and was able to offer frontend editing without too much
effort.