✨(courses) add a "code" field on the course model
We want to synchronize course runs of a course to a LMS's courses.
For this, we need a unique way of identifying courses. The code
should be unique across all courses (but this can't be done as a
database constraint because a course has a public counterpart and
may also have snapshots). We enforce uniqueness via tests in the
"validate_unique" method and make sure it is called on all saves.