✨(courses) add an API endpoint to synchronize course runs
We want to be able to update course runs from a LMS each time
a course's details are updated. This idea is to provide a hook
to which a course run's details can be posted and that will sort
it out whatever the situation: course existing or not, draft or
published... course run existing or not, etc.

We thought of offering 2 synchronization modes:
- sync to draft: only the draft version of the course run is updated
  and the related course is marked dirty,
- sync to public: both the draft and public course runs are updated

A manual mode is also provided to allow editing some course runs
manually in richie at the same time...