🐛(courses) fix displaying course runs in list on course detail page
The template code that was taking care of displaying the course run details
with or without a title, was broken in french because the translation included
a comma where it shouldn't. It turns out that the implementation is a git too
"smart" and thus fragile. We refactor it to use a good old "if" clause.