✨(plugins) add a "person" variant to the glimpse plugin
We want to be able to refer to a person without having to create a
person page. The glimpse plugin is perfect for this because it has
the required fields: avatar, name, bio and link.

I chose to make the glimpse plugin generate the exact same markup
as the person glimpse so that the css is entirely shared.