💚(ci) fix docker login run with forks
People which submit PR through a fork see ci job `docker-build` failed. For
security reason, we don't share our CI environment variables with forked PR.
This cause an issue on `docker login` because credential variables are not
exposed.

To fix it we test that docker credentials are exposed, if not process continues
in anonymous mode.