🐛(courses) show an error banner when template is misused
Page based on [model]_detail template must have a model attached, if not,
an 500 error occured. To prevent this error to appear, we manage the case where
no model is attached to the page and display an error banner to alert the user
that he/she is misusing the template.
It concerns 4 models: category, course, organization and person.

Issue:  #1280