✨(frontend) add fonzie authentication backend
A first step to integrate Richie with Joanie is to get a jwt token from the
identity provider. For now, identity provider is OpenEdx.
To retrieve authenticated user information with a signed jwt token, we added an
API route to OpenEdX through Fonzie. So, we have to add the support of this new
authentication route from Richie.