✨(blogposts) order blog posts by reverse publication date on list view
We were sorting blog posts as per the page tree to allow drag-and-drop
reordering. However, the DjangoCMS page tree does not scale to a big number
of pages and we deactivated digging into blog post pages via the admin on some
of our sites. As a result we needed to find a better way to order blog posts.
The publication date seems like the best criteria. It can be changed via the
toolbar on a given blog post.