🔧(sandbox) deactivate language fallback redirect in sandbox
This behavior is weird because, someone you is visiting the site in
a language will be permanently redirected to the other language as
soon as he/she encounters a page that requires a language fallback.
After this redirect, the user will need to go through the language
selector to go back to the language he/she was browsing.

What we want is to show the content in the language of the fallback
but continue on the selected language for subsequent pages.