♻️(courses) refactor plugins to handle language fallbacks
When a page extention plugin (CoursePlugin, OrganizationPlugin, etc.)
was pointing to a page that was not published in the current language,
but published in another language, we were displaying broken plugins.
To fix this, I implemented language fallback on the page extention
plugins, getting inspiration from the DjangoCMS main view where pages
are retrieved taking into account their language fallback.

I also took this opportunity to show plugins on the page only for published
pages. This decision follows a complaint from our support team stating that
"In its draft version, we want to see the page exactly how it will look
when we press the publish button".