🚸(plugins) improve empty label in LTI provider select box
We didn't want to add a specific choice in the "lti_provider_id" field
for custom configuration. But it is still possible to customize the
empty label to make it look like an option. In the database, we keep
the benefit of having the None and not a fake provider id when none is
selected...