⚡️(courses) cache getting course runs dict from the course models
On one of our sites, we want to display the first open course run
in the top header of the course syllabus, and only the remaining
ones in the side column.

This will be done via overidding template blocks and we will need to
call the method "get_course_runs_dict" several times in the template.
This method being very costly, it is a good idea to first make sure
that it is cached on the instance.