✨(courses) use lang fallback to get reverse related page extensions
When we rely on page extension plugins added in a language to display
on other languages via fallbacks, the reverse relation was broken.
For example: organizations added to the course page in english. They
were displayed on the course's french page but the course did not
display on the organization's french page.