♻️(frontend) do not make request for `isEnrolled` enrollment API method
In the case of OpenEdX LMS, within the enrollment API  `get` and `isEnrolled`
methods was the same request then `isEnrolled` just extracted the useful data to
know if user was enrolled or not to the provided resource_link. As `get` method
was already retrieve the enrollment data, we can pass these data to the
`isEnrolled` method to avoid an extra requests. In case of other LMS, if we need
to get these information from a request, it still totaly possible.