🐛(frontend) fix iframeResizer instantiation
The logic to retrieve the container of the LTIConsumer component did not support
the case when several plugins was on the same page. So even if there are several
plugins on the page, iframeResizer was only target the first one.

Furthermore, LTIConsumerPlugins styles was linked to the subheader container. So
if we put this plugin outside of this element, there was layout issue.