🐛(frontend) fix course teaser layout on safari
On Safari, when user launches course teaser in fullscreen mode some DOM elements
appear on the top of the video. It was due to a bug with Safari related to the
animation fill mode applied to the teaser container.

Also an offset was applied to iframe when its content loaded.