✅(search) clarify languages used on course unpublished in a language
When a course is unpublished in one language, but it has related
objects that are still published in this language, should or shouldn't
we use this language's content for the related objects when building the
course glimpse? For the moment it does.

This choice is controversial and I had to write 2 tests to fully
understand it. So I propose to commit them to keep this memory.

Note: We could argue in the future that a course glimpse should be built
with content only in the same language and not with "as few fallback
languages as possible, using available content in each language".