🔧(docker) enable Richie to run locally with ES6 or ES7 based on env
As we upgrade Richie to ElasticSearch 7, we don't want to force all
users to upgrade the exact moment we do. Instead, we want to support
older versions for as long as it is practical and useful.

The first step towards this is to make sure we can run the version
we want locally with Richie, even though we use the latest version (7)
by default.