👷(elasticsearch) update the CI to add tests for Elasticsearch 7
As we make sure Richie runs on both ES6 and ES7 right now, we need
to be careful of regressions as it would be easy for developers
running ES7 locally and in unit tests to break ES6 without realizing
it.

To prevent this, we need to add a CI job to automatically test ES6,
that we'll keep around for as long as Richie officially supports it.