⬆️(indices) update index manager broken index handling
Richie operates with somewhat light indices that can be regenerated
often, to avoid the hassle of maintaining a permanent database in
addition to PostgreSQL.

In these semi frequent index regeneration operations, we encountered
issues where a broken index could stop the whole process and needed
manual intervention to repair the cluster. We implemented a mechanism
to catch these errors and handle them. This relied on having the name
of the broken index provided in the error.

Unfortunately, ES7 no longer cleanly provides the broken index' name,
like ES6. We therefore added a secondary fallback mechanism where
we extract the index name from an error message to automatically
handle the error.