✨(api) order non-course search results alphabetically
As we introduced compatibility with both ES6 & ES7 in Richie, we
encountered a case where ordering of objects was random when there
is no search query to score them on (except for courses which have
a script).

To fix this (which broke our compatibility tests) and as a feature
we already have in our backlog, we ordered categories, organizations
and persons alphabetically in search responses.