⬆️(elasticsearch) adapt to the ES7 results count format
ElasticSearch changed the way it passes information about the
total number of hits for a search. The integer nested in the response
object was replaced with a dict that uses a value and a relation
key to communicate the same information (for our use cases).

Update Richie to the way ES7 works, and use the compatibility layer
later on to patch the ES6 response so our code does not break.