🐛(website) override IS_PULL_REQUEST to force Docusaurus deploy
Docusaurus only deploys the website if the CI_PULL_REQUEST and
CIRCLE_PULL_REQUEST environment variables are both null. This is
apparently not the case anymore on a build of the master branch so
we decide to force it on the command.

We can not force deployment with the FORCE_DEPLOY environment
variable because it was not released in Docusaurus v1.