✨(search) improve aggregations ordering on filters
Facet counts (aggregations) in the filters on course searches had
different logics apply to them for ordering depending on the kind
of filter definition.

We rationalized the way facets are ordered to be either "conf",
"count" then "alphabetical" or "alphabetical" then "count" so we
always have a consistent expected ordering of results. It also
makes it way easier to write tests as there is only one correct
ordering.