🐛(search) lazily retrieve es version
Retrieve the elasticsearch server version on initialization could creates bug in
contexts where elasticsearch host is not accessible. e.g: during a docker build
if we run a manage command, an elasticsearch client is initiliazed and build
failed due to a ConnectionError. So a workaround is to lazily retrieve the
elasticsearch version the first time we need to know the version then reuse
the value.