♻️(search) stop relying on MPTT paths to define search filters
Filters based on the IndexableMPTTFilterDefinition, which are filters
defined by a base page and for which a hierarchy of pages indexed in
ES is the list of available filters, were using a regex filter that
required the MPTT path of said filters to be used as IDs and indexed
on courses.

We replaced the regexp with a (cached) database call to directly
provide a list of children instead of an MPTT based regex. This
will enable us to link filters with courses through proper IDs as
filters do not rely on MPTT paths in their definition.