💄(subheader) use a dedicated variable to set aside column width
Previously, we use `$r-course-aside` variable to set both the course content
aside column width and the subheader aside column width. But in some use cases,
we want to define this variable to `null` to stretch `course__content` in full
width. But we do not want this behavior for the subheader block. So we decided
to use a dedicated variable to set the subheader aside column width.