♻️(api) replace regex-based _include params with list-based _aggs
The course search API supports regex-based [filter]_include parameters
that allow requests to specify which terms they want to include in
facets for a given filter. These regexes are intended to match on the
MPTT paths for related pages, used as IDs.

As we want to get rid of those as IDs, we need to rework those query
string parameters to use another mechanism. Both use cases currently
supported in the Richie client can be replaced with a list of the
requested terms.