✨(api) add _children_aggs params for indexable filters
As we continue to work toward removing page MPTT paths as IDs for
elements in Richie's ES indices, we're covering another use case for
MPTT based filtering: a valid client-side use case is to retrieve
facets for all children of a given filter, so that they can be
displayed when the user clicks to "open" the parent.

As the frontend does not have a knowledge of all the children of
any filter value, we cannot use the regular [filter]_aggs param
here. Instead, we're adding a new [filter]_children_aggs which
leaves the work of finding IDs for the children to the backend.