✨(forms) update array field to accept comma-separated lists
As we update our API to accept lists of IDs of elements that
allow API clients to specify aggregations, we want to avoid making
URLs too long by repeating the key for each ID.

On the frontend, we can use our library's options to just send a
comma-separated list of IDs. On the backend side, we need to handle
these lists in our ArrayField.