💄(frontend) make scheme svg color more easily overridable
If the selected scheme has a `svg-color` property, `r-scheme-colors` applied
this color to all svg children of the selected element. Coupled with the use of
`detail-block` mixin, in some case, the css selector in charge to style svg had
a too high priority selector and we were unable to change svg color from a lower
level. So we decide to enlight as possible selector in charge to apply
svg-color, to do that we refactor `detail-block` and `r-scheme-colors` mixins.