♻️(models) add a mixin to get the ES ID of an indexable object
All our indexable objects use the public extension's related
page id, as a string, as the object ID in ElasticSearch. This is
a little annoying to poke every time for every object.

We added a mixin with a shared method that produces the ES ID,
no matter if it is called from a public extension or a draft
extension.