♻️(backend) replace all MPTT-path-as-ID uses with actual IDs
As the final part of our refactor, now that no part of Partaj
relies on organizations & categories IDs being MPTT paths
masquerading as IDs, we can replace the paths with proper IDs for
the same objects.

Mimicking what was done for persons, we used the ID for the
published page linked to the object. To have filters and facets
match, we needed to put those IDs both on the courses and on the
categories and organizations.