🚑️(frontend) fix utf-16 to base64 encoding error
Since DOMStrings are 16-bit-encoded strings, in most
browsers calling window.btoa on a Unicode string will cause a Character Out Of
Range exception if a character exceeds the range of a 8-bit byte (0x00~0xFF).
So we have to use a custom base 64 parser to encode and decode UTF-16 content
without error.