🐛(search) handle too-short query strings in <Search />
When the user makes use of our <Search /> UI, if they type a textual
query that is too short to be used to search, we manage it by not
triggering a search, which users understand as a nudge to keep typing.

There are cases, however, when a user can land on the course search
page with a too-short query string in the URL, such as when they
typed that short query string in the search bar in the homepage or
in the header.

In that case, we want to display search results that do not make use
of their short query, as we would have done if they just typed 2
characters in the search view directly. Unlike that case however,
we need to warn them that their query was ignored as the interface
did not give them a hint by its state while they were typing.