💄(frontend) stack organization logo to the right
To improve course glimpse layout, we decided to stack the organization logo
to the right of the course glimpse tile.

Furthermore, in some case where category label is long, organization logo was
above this element. We propose to move organization logo behind this label to
guarantee that label are always fully readable.