🐛(api) handle malformed autocomplete errors with 400 errors
Malformed autocomplete requests to the search endpoints, which did
not contain a "query" parameter, caused the server to throw a 500
error. Those requests should not exist in the first place, but the
API should also not error out with a 500. Instead, we'd rather
handle the error and return a 400 error with an explanation.