♻️(frontend) handle 400 errors on autocomplete calls
We just improved the autocomplete API by handling incorrect requests
with a 400 error with an explanation, instead of raising an exception
and returning a 500.

To take better care of the source of these malformed requests, we
want to catch those errors on the frontend side and return them with
their context so we can find out their cause.