✨(search) update ES indices when a page is moved
When a page is moved in the MPTT tree, it can change the paths
of an unbounded number of other pages linked to the same kind
of objects. Right now, paths are only used for categories.

Therefore, when a category is moved in the page tree, we need to
re-index all categories, as we do not know which pages may be
impacted by the move.