📌(backend) pin redis-py until fixed 4+ version is ready
Newer versions of redis-py, which are automatically pulled into
Richie through our dependency on django-redis, are broken (see
https://github.com/redis/redis-py/issues/1714). This will be
fixed in the next version 4.0.2.

django-redis pinned the version of redis-py on master but did not
issue a release with this dependency fix. We can do the same
pinning as them (`redis>=3,<4`) to stop the breaking for now.