🔧(cache) define default value for maximum browsers cache
If the maximum browser cache is not set, we fall victim to a bug in
DjangoCM which is that the browser cache and server cache are set
to the same value. It means that we can not set a long server cache
(which we want since it is invalidated as soon as a change is published)
without setting a long browser cache at the same time (which we don't
want because we want changes to be effective rapidly).